file renamed to websocket-server and tests fixed
