+++ lib/testsuite/selector-playground/index.js
@@ -3,7 +3,7 @@ const {crxfile} = require('@nightwatch/nightwatch-inspector');
 const stripAnsi = require('strip-ansi');
 const Debuggability = require('../../utils/debuggability');
 const Utils = require('../../utils');
-const WebSocket = require('./playground-listener');
+const WebSocket = require('./websocket-server');
 const {Logger} = Utils;
 
 module.exports = class SelectorPlaygroundServer extends WebSocket {
